+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+* removed --search-redhat on service products-contain-component
+* several minor output fixes
+
 ## [0.5.1] - 2024-01-19
 ### Changed
 * --filter-rh-naming enabled by default (GRIF-121)